split feature(s) domain module
