+++ bofire/benchmarks/benchmark.py
@@ -8,7 +8,7 @@ from multiprocess.pool import Pool
 from tqdm import tqdm
 
 from bofire.domain.domain import Domain
-from bofire.domain.features import OutputFeature
+from bofire.domain.feature import OutputFeature
 from bofire.domain.objectives import Objective
 from bofire.strategies.strategy import Strategy
