+++ bofire/benchmarks/multi.py
@@ -12,14 +12,13 @@ from scipy.special import gamma
 
 from bofire.benchmarks.benchmark import Benchmark
 from bofire.domain import Domain
-from bofire.domain.features import (
+from bofire.domain.feature import (
     CategoricalDescriptorInput,
     ContinuousInput,
     ContinuousOutput,
     InputFeature,
-    InputFeatures,
-    OutputFeatures,
 )
+from bofire.domain.features import InputFeatures, OutputFeatures
 from bofire.domain.objectives import (
     MaximizeObjective,
     MaximizeSigmoidObjective,
