+++ bofire/models/model.py
@@ -6,7 +6,8 @@ import pandas as pd
 from pydantic import Field, validator
 from sklearn.model_selection import KFold
 
-from bofire.domain.features import InputFeatures, OutputFeatures, TInputTransformSpecs
+from bofire.domain.feature import TInputTransformSpecs
+from bofire.domain.features import InputFeatures, OutputFeatures
 from bofire.domain.util import PydanticBaseModel
 from bofire.models.diagnostics import CvResult, CvResults
 from bofire.utils.enum import OutputFilteringEnum
