+++ bofire/strategies/botorch/qehvi.py
@@ -17,7 +17,7 @@ from botorch.utils.multi_objective.box_decompositions.non_dominated import (
 from pydantic import confloat
 
 from bofire.domain.constraint import Constraint, NChooseKConstraint
-from bofire.domain.features import Feature
+from bofire.domain.feature import Feature
 from bofire.domain.objectives import (
     BotorchConstrainedObjective,
     MaximizeObjective,
