+++ bofire/strategies/strategy.py
@@ -8,7 +8,7 @@ from pydantic.types import NonNegativeInt, PositiveInt
 
 from bofire.domain.constraints import Constraint
 from bofire.domain.domain import Domain
-from bofire.domain.features import Feature, OutputFeature, TInputTransformSpecs
+from bofire.domain.feature import Feature, OutputFeature, TInputTransformSpecs
 from bofire.domain.objectives import Objective
