+++ tests/bofire/conftest.py
@@ -1,7 +1,7 @@
 from pytest import fixture
 
 from bofire.domain.constraints import Constraint
-from bofire.domain.features import Feature
+from bofire.domain.feature import Feature
 from bofire.domain.objectives import Objective
 from tests.bofire import specs
