+++ tests/bofire/domain/test_domain_validators.py
@@ -8,7 +8,7 @@ import pytest
 
 from bofire.domain.constraint import LinearEqualityConstraint
 from bofire.domain.domain import Domain
-from bofire.domain.features import (
+from bofire.domain.feature import (
     CategoricalDescriptorInput,
     CategoricalInput,
     ContinuousInput,
