+++ tests/bofire/models/gps/test_gps.py
@@ -8,13 +8,8 @@ from botorch.models.transforms.input import (
 )
 from botorch.models.transforms.outcome import Standardize
 
-from bofire.domain.features import (
-    CategoricalInput,
-    ContinuousInput,
-    ContinuousOutput,
-    InputFeatures,
-    OutputFeatures,
-)
+from bofire.domain.feature import CategoricalInput, ContinuousInput, ContinuousOutput
+from bofire.domain.features import InputFeatures, OutputFeatures
 from bofire.models.gps import MixedSingleTaskGPModel, SingleTaskGPModel
 from bofire.models.gps.kernels import HammondDistanceKernel, RBFKernel, ScaleKernel
 from bofire.utils.enum import CategoricalEncodingEnum, ScalerEnum
