+++ tests/bofire/models/test_feature_importance.py
@@ -1,12 +1,8 @@
 import pandas as pd
 import pytest
 
-from bofire.domain.features import (
-    ContinuousInput,
-    ContinuousOutput,
-    InputFeatures,
-    OutputFeatures,
-)
+from bofire.domain.feature import ContinuousInput, ContinuousOutput
+from bofire.domain.features import InputFeatures, OutputFeatures
 from bofire.models.diagnostics import metrics
 from bofire.models.feature_importance import (
     combine_permutation_importances,
