+++ tests/bofire/strategies/botorch/test_qehvi.py
@@ -12,7 +12,7 @@ from botorch.acquisition.multi_objective.objective import WeightedMCMultiOutputO
 
 from bofire.benchmarks.multi import C2DTLZ2, DTLZ2
 from bofire.domain.domain import Domain
-from bofire.domain.features import ContinuousInput, ContinuousOutput
+from bofire.domain.feature import ContinuousInput, ContinuousOutput
 from bofire.domain.objectives import MaximizeObjective, MinimizeObjective
 from bofire.samplers import PolytopeSampler
 from bofire.strategies.botorch.qehvi import BoTorchQehviStrategy, BoTorchQnehviStrategy
