+++ tests/bofire/utils/test_reduce.py
@@ -2,11 +2,11 @@ import numpy as np
 import pytest
 
 from bofire.domain import Domain
-from bofire.domain.constraints import (
+from bofire.domain.constraint import (
     LinearEqualityConstraint,
     LinearInequalityConstraint,
 )
-from bofire.domain.features import CategoricalInput, ContinuousInput, ContinuousOutput
+from bofire.domain.feature import CategoricalInput, ContinuousInput, ContinuousOutput
 from bofire.utils.reduce import (
     check_domain_for_reduction,
     check_existence_of_solution,
