+++ tests/bofire/utils/test_torch_tools.py
@@ -3,16 +3,12 @@ import pytest
 import torch
 
 from bofire.domain import Domain
-from bofire.domain.constraints import (
+from bofire.domain.constraint import (
     LinearEqualityConstraint,
     LinearInequalityConstraint,
 )
-from bofire.domain.features import (
-    CategoricalInput,
-    ContinuousInput,
-    ContinuousOutput,
-    OutputFeatures,
-)
+from bofire.domain.feature import CategoricalInput, ContinuousInput, ContinuousOutput
+from bofire.domain.features import OutputFeatures
 from bofire.domain.objectives import (
     MaximizeObjective,
     MaximizeSigmoidObjective,
